+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## [0.19.0] - 2022-02-02
+### Added
+- Support for Angular client with `--name` option
+- Added test cases for Angular client
+
 ## [0.18.2] - 2022-02-02
 ### Fixed
 - Updated dependencies